:construction: Working on image particles

+++ Engine/Vector.py
@@ -1,6 +1,6 @@
 
 class Vector2D():
-    def __init__(self, x: int = 0, y: int = 0):
+    def __init__(self, x: float = 0, y: float = 0):
         self.x = x
         self.y = y
